feat(markdown-toolbar): implement markdown formatting toolbar for comment and reply inputs
